[ENHANCEMENT] Better chart names
Grouped bar charts of ggplots uses the names of the variables as names of each axis on the chart.

On this commit we use more descriptive variable names in order to get better aesthetic on the chart.

Also, solved an annoying warning about the fastas not being completes. Just by adding a line break at the end of every fasta file.